Casts to match correct SString constructor now there are int and size_t versions.

+++ gtk/SciTEGTK.cxx
@@ -2291,7 +2291,7 @@ void SciTEGTK::TabSizeDialog() {
 	GtkWidget *labelTabSize = TranslatedLabel("_Tab Size:");
 	table.Label(labelTabSize);
 
-	SString tabSize(wEditor.Call(SCI_GETTABWIDTH));
+	SString tabSize(static_cast<int>(wEditor.Call(SCI_GETTABWIDTH)));
 	dlgTabSize.entryTabSize.Create(tabSize.c_str());
 	table.Add(dlgTabSize.entryTabSize);
 	dlgTabSize.entryTabSize.ActivatesDefault();
@@ -2301,7 +2301,7 @@ void SciTEGTK::TabSizeDialog() {
 	GtkWidget *labelIndentSize = TranslatedLabel("_Indent Size:");
 	table.Label(labelIndentSize);
 
-	SString indentSize(wEditor.Call(SCI_GETINDENT));
+	SString indentSize(static_cast<int>(wEditor.Call(SCI_GETINDENT)));
 	dlgTabSize.entryIndentSize.Create(indentSize.c_str());
 	table.Add(dlgTabSize.entryIndentSize);
 	dlgTabSize.entryIndentSize.ActivatesDefault();
